上传排序、列表默认排序、根据角色自营、商品删除、是否发售默认

+++ src/main/java/cn/offway/athena/service/impl/PhRoleServiceImpl.java
@@ -113,14 +113,16 @@ public class PhRoleServiceImpl implements PhRoleService {
 		
 		Long phRoleId = phRole.getId();
 		
-		List<Long> idList = (List<Long>)ListUtils.toList(resourceIds);
-		List<PhResource> phResources = phResourceRepository.findByIds(idList);
-		List<PhRoleresource> roleresources = new ArrayList<>();
-		for (PhResource phResource : phResources) {
-			getResources(now, phRoleId, roleresources, phResource);
+		if(null != resourceIds){
+			List<Long> idList = (List<Long>)ListUtils.toList(resourceIds);
+			List<PhResource> phResources = phResourceRepository.findByIds(idList);
+			List<PhRoleresource> roleresources = new ArrayList<>();
+			for (PhResource phResource : phResources) {
+				getResources(now, phRoleId, roleresources, phResource);
+			}
+			phRoleresourceRepository.deleteByRoleId(phRoleId);
+			phRoleresourceRepository.save(roleresources);
 		}
-		phRoleresourceRepository.deleteByRoleId(phRoleId);
-		phRoleresourceRepository.save(roleresources);
 	}
 
 	/**
